linux: fix local SSD setup on gcp amd instances (bug 1990389)
The AMD machine types (e.g. c3d) appear to expose the persistent drive
as nvme, so the assumptions in the ephemeral-disks setup script are
wrong: we can end up with nvme0n1 and nvme1n1 disks, one of which is
persistent and the other is local.  We assumed there'd be a single
controller, and nvme0n* would be local devices, so we'd entirely miss
nvme1n1, and proceed just using the persistent disk, which is slow and
small.

Instead, use the symlinks in /dev/disk/by-id to find the local SSDs in a
more robust way.

Refs:
https://cloud.google.com/compute/docs/disks/disk-symlinks
https://cloud.google.com/compute/docs/disks/local-ssd#device_names